Read Desc
changed the location of dbconnect.php. It is now in the main directory. Also changed the location of pagination.php to be in the main directory (because it is used so much).
You must delete your current database (named "stuco) and create a new one named "stuco". Then select import and upload the database.sql file.
In addition made all of the created tasks create their own new page.